+++ ubiquitous-asset-management/src/main/java/com/github/jacobbishopxy/ubiquitousassetmanagement/portfolio/models/Constituent.java
@@ -24,7 +24,7 @@ public class Constituent {
   @Id
   @Column(columnDefinition = "serial")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
   @ManyToOne(fetch = FetchType.EAGER)
   @JoinColumn(name = "portfolio_adjustment_record_id")
@@ -125,19 +125,19 @@ public class Constituent {
   // Accessors
   // =======================================================================
 
-  public int getAdjRecordId() {
-    Integer id = adjustmentRecord.getId();
+  public Long getAdjRecordId() {
+    Long id = adjustmentRecord.getId();
     if (id == null) {
       throw new IllegalArgumentException("The adjustment record id cannot be null.");
     }
     return id;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
